# control-tower/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/schedule": {
            "get": {
                "description": "gets schedule by the user email obtained in the JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SCHEDULES"
                ],
                "summary": "get schedules by creator.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Zero indexed",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PaginatedNotifications"
                        }
                    }
                }
            },
            "post": {
                "description": "create schedule.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "SCHEDULES"
                ],
                "summary": "create schedule.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.NotificationInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/schedule/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SCHEDULES"
                ],
                "summary": "get schedules by creator.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "schedule ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PaginatedNotifications"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a schedule from both the scheduler service and db.",
                "tags": [
                    "SCHEDULES"
                ],
                "summary": "deletes a schedule from both the scheduler service and db.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "patch existing schedule by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SCHEDULES"
                ],
                "summary": "patch existing schedule by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.PatchNotification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Notification"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Notification": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/service.ClientDto"
                },
                "createdBy": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deliveryMethods": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controller.PaginatedNotifications": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Notification"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "service.ClientDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lastUpdateAt": {
                    "type": "string"
                },
                "optIn": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "service.Notification": {
            "type": "object",
            "required": [
                "clientId",
                "createdBy",
                "date",
                "deliveryMethods",
                "id",
                "status"
            ],
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "clientToken": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deliveryMethods": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "service.NotificationInput": {
            "type": "object",
            "required": [
                "clientId",
                "date",
                "deliveryMethods"
            ],
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deliveryMethods": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "service.PatchNotification": {
            "type": "object",
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deliveryMethods": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}